Standardise import shared example to receive a key only and add negative
All the option / flag / boolean tests accept an options hash key as a symbol.  For some reason I created the import option version of the test to accept a hash.  Correcting this so it brings the test in line with the others.

Also added the negative test in the option test to align it with the common version.